@@ -1,5 +1,4 @@
-import { Subject } from "rxjs/Subject";
-import { Observable } from "rxjs/Observable";
+import { Subject, Observable } from "rxjs";
 
 import { injectable } from "app/di";
 import { ErrorDef } from "app/utils/error-helpers";
