@@ -1,4 +1,4 @@
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 
 import { LoginStatus } from "app/services/login/current-user.service";
 
