@@ -1,12 +1,9 @@
 import * as React from "react";
-import * as cn from "classnames";
 
 import { $T } from "app/i18n/strings";
 
 import { ModalComponent } from "app/ui/_generic/modal.component";
-import { ProgramModel } from "app/services/program/program.model";
 import { AlertMessageComponent } from "app/ui/_generic/alert-message.component";
-import { NoDataComponent } from "app/ui/_generic/no-data.component";
 
 interface IComponentState {
   errorMessage: string;
@@ -14,7 +11,7 @@ interface IComponentState {
 }
 
 interface IComponentProps {
-  program: ProgramModel;
+  programName: string;
   onDelete(): Promise<void>;
   onClose(): void;
 }
@@ -30,7 +27,6 @@ export class DeleteProgramModalComponent extends React.Component<IComponentProps
   }
 
   render(): JSX.Element | null {
-    const p = this.props.program;
     return (
       <ModalComponent
         show
@@ -50,34 +46,8 @@ export class DeleteProgramModalComponent extends React.Component<IComponentProps
               readOnly
               className="input"
               id="program-name-in-delete-dialog"
-              value={this.props.program.name}
+              value={this.props.programName}
             />
-          </div>
-        </div>
-        <div className="field">
-          <label className="label">{$T.gallery.editedDate}</label>
-          <div className="control">
-            <input
-              type="text"
-              readOnly
-              className="input"
-              id="program-edited-in-delete-dialog"
-              value={new Date(this.props.program.dateLastEdited).toLocaleDateString()}
-            />
-          </div>
-        </div>
-        <div className="field">
-          <label className="label">{$T.gallery.galleryImage}</label>
-          <div className="has-text-centered">
-            <div className="box is-inline-block">
-              {this.props.program.screenshot ? (
-                <figure className="has-text-centered">
-                  <img src={this.props.program.screenshot} />
-                </figure>
-              ) : (
-                <NoDataComponent iconClass="fa-picture-o" title={$T.gallery.noImage} />
-              )}
-            </div>
           </div>
         </div>
       </ModalComponent>
