@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Subscription } from "rxjs/Subscription";
+import { Subscription } from "rxjs";
 
 import { resolveInject } from "app/di";
 import { DependecyInjectionSetupService } from "app/di-setup";
@@ -52,8 +52,11 @@ export class SignInStatusComponent extends React.Component<IComponentProps, ICom
             <i className="fa fa-google" aria-hidden="true" />&nbsp;
           </>
         );
+      case AuthProvider.none:
+        return <></>;
+      default:
+        throw new Error("Auth provider is not known");
     }
-    return null;
   }
 
   render(): JSX.Element {
