@@ -1,44 +1,49 @@
 import * as React from "react";
 import { RouteComponentProps } from "react-router-dom";
-import { Subject } from "rxjs/Subject";
-import { BehaviorSubject } from "rxjs/BehaviorSubject";
+import { Subject } from "rxjs";
 
-import { callActionSafe, ErrorDef } from "app/utils/error-helpers";
-import { as } from "app/utils/syntax-helpers";
+import { callActionSafe } from "app/utils/error-helpers";
 import { resolveInject } from "app/di";
 import { Routes } from "app/routes";
 import { $T } from "app/i18n/strings";
 import { checkIsMobileDevice } from "app/utils/device-helper";
-import { NotificationService } from "app/services/infrastructure/notification.service";
 import { ErrorService } from "app/services/infrastructure/error.service";
 import { TitleService } from "app/services/infrastructure/title.service";
 import { NavigationService } from "app/services/infrastructure/navigation.service";
 import { IUserSettingsService, IUserSettings } from "app/services/customizations/user-settings.service";
-import { ProgramModel } from "app/services/program/program.model";
-import { ProgramModelConverter } from "app/services/program/program-model.converter";
 import { ThemesService, Theme } from "app/services/customizations/themes.service";
 import { TurtlesService } from "app/services/customizations/turtles.service";
 import { ProgramExecutionContext } from "app/services/program/program-execution.context";
-import { TutorialsContentService, ITutorialInfo } from "app/services/tutorials/tutorials-content-service";
+import {
+  TutorialsContentService,
+  ITutorialInfo,
+  ITutorialStepContent
+} from "app/services/tutorials/tutorials-content-service";
 import { EventsTrackingService, EventAction } from "app/services/infrastructure/events-tracking.service";
 import { tutorialsDefaultLayout, tutorialsDefaultMobileLayout } from "app/ui/tutorials/tutorials-default-goldenlayout";
 import { MainMenuComponent } from "app/ui/main-menu.component";
-import { GoldenLayoutComponent, IPanelConfig } from "app/ui/_shared/golden-layout.component";
-import { CodePanelComponent, ICodePanelComponentProps } from "app/ui/playground/code-panel.component";
-import { OutputPanelComponent, IOutputPanelComponentProps } from "app/ui/playground/output-panel.component";
-import { TutorialViewComponent, ITutorialViewComponentProps } from "app/ui/tutorials/tutorial-view.component";
 import { LoadingComponent } from "app/ui/_generic/loading.component";
+import { ReactGoldenLayout } from "app/ui/_generic/react-golden-layout/react-golden-layout";
+import { ReactGoldenLayoutPanel } from "app/ui/_generic/react-golden-layout/react-golden-layout-panel";
+import { CodePanelComponent } from "app/ui/code-panel/code-panel.component";
+import { TutorialViewComponent } from "app/ui/tutorials/tutorial-view.component";
+import { OutputPanelComponent } from "app/ui/output-panel/output-panel.component";
+
+import "./tutorials.page.component.less";
 
 interface IComponentState {
   isLoading: boolean;
-  userSettings?: IUserSettings;
-  theme?: Theme;
-  turtleImage?: HTMLImageElement;
+
   tutorials?: ITutorialInfo[];
   tutorialId?: string;
   stepId?: string;
-  program?: ProgramModel;
-  pageLayoutConfigJSON?: string;
+
+  programCode: string;
+
+  userSettings?: IUserSettings;
+  theme?: Theme;
+  turtleImage?: HTMLImageElement;
+  layoutLocalStorageKey?: string;
 }
 
 interface IComponentProps extends RouteComponentProps<ITutorialPageRouteParams> {}
@@ -49,7 +54,6 @@ export interface ITutorialPageRouteParams {
 }
 
 export class TutorialsPageComponent extends React.Component<IComponentProps, IComponentState> {
-  private notificationService = resolveInject(NotificationService);
   private navService = resolveInject(NavigationService);
   private errorService = resolveInject(ErrorService);
   private titleService = resolveInject(TitleService);
@@ -60,9 +64,8 @@ export class TutorialsPageComponent extends React.Component<IComponentProps, ICo
   private eventsTracking = resolveInject(EventsTrackingService);
 
   private isMobileDevice = checkIsMobileDevice();
-  private executionService = new ProgramExecutionContext();
-  private codeChangesStream = new Subject<string>();
-  private layoutChangesStream = new Subject<void>();
+  private executionContext = new ProgramExecutionContext();
+  private resizeEventsSubject = new Subject<void>();
 
   private defaultLayoutConfigJSON = JSON.stringify(
     this.isMobileDevice ? tutorialsDefaultMobileLayout : tutorialsDefaultLayout
@@ -71,7 +74,8 @@ export class TutorialsPageComponent extends React.Component<IComponentProps, ICo
   constructor(props: IComponentProps) {
     super(props);
     this.state = {
-      isLoading: true
+      isLoading: true,
+      programCode: ""
     };
   }
 
@@ -81,16 +85,18 @@ export class TutorialsPageComponent extends React.Component<IComponentProps, ICo
     await this.loadData(this.props);
   }
 
+  async componentWillReceiveProps(newProps: IComponentProps) {
+    if (
+      newProps.match.params.tutorialId != this.state.tutorialId ||
+      newProps.match.params.stepId != this.state.stepId
+    ) {
+      await this.loadData(newProps);
+    }
+  }
+
   componentWillUnmount() {
     /** */
   }
-
-  layoutChanged = async (newLayoutJSON: string) => {
-    await this.userSettingsService.update(
-      this.isMobileDevice ? { tutorialsLayoutMobileJSON: newLayoutJSON } : { tutorialsLayoutJSON: newLayoutJSON }
-    );
-    this.layoutChangesStream.next();
-  };
 
   async loadData(props: IComponentProps) {
     this.setState({
@@ -112,8 +118,6 @@ export class TutorialsPageComponent extends React.Component<IComponentProps, ICo
       return;
     }
 
-    const program = ProgramModelConverter.createNewProgram(undefined, "", "", "");
-
     let tutorialId = props.match.params.tutorialId;
     let stepId = props.match.params.stepId;
     if (!tutorialId || !stepId) {
@@ -132,116 +136,105 @@ export class TutorialsPageComponent extends React.Component<IComponentProps, ICo
       turtleImage,
       theme,
       tutorials,
-      program,
+      programCode: "",
       tutorialId,
       stepId,
-      pageLayoutConfigJSON: this.isMobileDevice
-        ? userSettings.tutorialsLayoutMobileJSON
-        : userSettings.tutorialsLayoutJSON
+      layoutLocalStorageKey:
+        this.userSettingsService.userSettingsKey + ":tutorials-layout" + (this.isMobileDevice ? "-mobile" : "-desktop")
     });
   }
 
-  onFixTheCode = (code: string) => {
-    if (this.state.program) {
-      this.codeChangesStream.next(code);
-      this.eventsTracking.sendEvent(EventAction.tutorialsFixTheCode);
+  handleLoadedTutorial = async (content: ITutorialStepContent) => {
+    if (content.initialCode) {
+      this.setState({ programCode: content.initialCode });
+      this.executionContext.executeProgram(content.initialCode);
     }
   };
 
-  onLoadedTutorial = async (tutorialId: string, stepId: string, initCode: string) => {
-    if (this.state.program) {
-      this.codeChangesStream.next(initCode);
-      this.executionService.executeProgram(initCode);
-    }
-
+  handleNavigationRequest = async (tutorialId: string, stepId: string) => {
     await this.userSettingsService.update({
-      currentTutorialInfo: {
-        tutorialId,
-        stepId
-      }
+      currentTutorialInfo: { tutorialId, stepId }
     });
-
     this.navService.navigate({
-      route: Routes.tutorialSpecified.build({
-        tutorialId: tutorialId,
-        stepId: stepId
-      })
+      route: Routes.tutorialSpecified.build({ tutorialId, stepId })
     });
+  };
+
+  fixTheCode = (code: string) => {
+    this.setState({ programCode: code });
+    this.executionContext.executeProgram(code);
+    this.eventsTracking.sendEvent(EventAction.tutorialsFixTheCode);
   };
 
   render(): JSX.Element {
     return (
       <div className="ex-page-container">
         <MainMenuComponent />
-        <div className="ex-page-content">
+        <div className="ex-page-content tutorials-page-component">
           {this.state.isLoading && (
-            <div className="golden-layout-component">
-              <div className="ex-page-content lm_content">
-                <LoadingComponent isLoading />
-              </div>
+            <div className="main-loading-container">
+              <LoadingComponent isLoading />
             </div>
           )}
 
-          {!this.state.isLoading &&
-            this.state.userSettings &&
+          {this.state.userSettings &&
             this.state.theme &&
             this.state.tutorials &&
             this.state.tutorialId &&
-            this.state.stepId &&
-            this.state.program && (
-              <GoldenLayoutComponent
-                initialLayoutConfigJSON={this.state.pageLayoutConfigJSON}
+            this.state.stepId && (
+              <ReactGoldenLayout
+                className="golden-layout-container"
+                configLayoutOverride={{
+                  settings: {
+                    showMaximiseIcon: false,
+                    showPopoutIcon: false,
+                    showCloseIcon: false
+                  },
+                  dimensions: { headerHeight: 32 }
+                }}
+                layoutLocalStorageKey={this.state.layoutLocalStorageKey}
                 defaultLayoutConfigJSON={this.defaultLayoutConfigJSON}
-                onLayoutChange={this.layoutChanged}
-                panelsReloadCheck={() => false}
-                panels={[
-                  as<IPanelConfig<TutorialViewComponent, ITutorialViewComponentProps>>({
-                    title: new BehaviorSubject(
-                      `<i class="fa fa-graduation-cap" aria-hidden="true"></i> ` + $T.tutorial.tutorialsTitle
-                    ),
-                    componentName: "tutorial-panel",
-                    componentType: TutorialViewComponent,
-                    props: {
-                      onFixTheCode: this.onFixTheCode,
-                      onLoadedTutorial: this.onLoadedTutorial,
-                      initialStepId: this.state.stepId,
-                      initialTutorialId: this.state.tutorialId,
-                      tutorials: this.state.tutorials
-                    }
-                  }),
-                  as<IPanelConfig<CodePanelComponent, ICodePanelComponentProps>>({
-                    title: new BehaviorSubject(`<i class="fa fa-code" aria-hidden="true"></i> ` + $T.program.code),
-                    componentName: "code-panel",
-                    componentType: CodePanelComponent,
-                    props: {
-                      editorTheme: this.state.theme.codeEditorThemeName,
-                      executionService: this.executionService,
-                      program: this.state.program,
-                      saveCurrentEnabled: false,
-                      externalCodeChanges: this.codeChangesStream,
-                      containerResized: this.layoutChangesStream
-                    }
-                  }),
-                  as<IPanelConfig<OutputPanelComponent, IOutputPanelComponentProps>>({
-                    title: new BehaviorSubject(
-                      `<i class="fa fa-television" aria-hidden="true"></i> ` + $T.program.output
-                    ),
-                    componentName: "output-panel",
-                    componentType: OutputPanelComponent,
-                    props: {
-                      logoExecutorProps: {
-                        runCommands: this.executionService.runCommands,
-                        stopCommands: this.executionService.stopCommands,
-                        makeScreenshotCommands: this.executionService.makeScreenshotCommands,
-                        onIsRunningChanged: this.executionService.onIsRunningChanged,
-                        isDarkTheme: this.state.theme.isDark,
-                        customTurtleImage: this.state.turtleImage,
-                        customTurtleSize: this.state.userSettings.turtleSize
-                      }
-                    }
-                  })
-                ]}
-              />
+                onLayoutChange={() => this.resizeEventsSubject.next()}
+              >
+                <ReactGoldenLayoutPanel
+                  id="code-panel"
+                  title={`<i class="fa fa-code" aria-hidden="true"></i> ${$T.program.codePanelTitle}`}
+                >
+                  <CodePanelComponent
+                    programName="Tutorial"
+                    programCode={this.state.programCode}
+                    hasChanges={false}
+                    executionContext={this.executionContext}
+                    onCodeChange={newCode => this.setState({ programCode: newCode })}
+                    editorTheme={this.state.theme.codeEditorThemeName}
+                    resizeEvents={this.resizeEventsSubject}
+                  />
+                </ReactGoldenLayoutPanel>
+                <ReactGoldenLayoutPanel
+                  id="output-panel"
+                  title={`<i class="fa fa-television" aria-hidden="true"></i> ${$T.program.outputPanelTitle}`}
+                >
+                  <OutputPanelComponent
+                    executionContext={this.executionContext}
+                    isDarkTheme={this.state.theme.isDark}
+                    turtleImage={this.state.turtleImage}
+                    turtleSize={this.state.userSettings.turtleSize}
+                  />
+                </ReactGoldenLayoutPanel>
+                <ReactGoldenLayoutPanel
+                  id="tutorial-panel"
+                  title={`<i class="fa fa-graduation-cap" aria-hidden="true"></i> ${$T.tutorial.tutorialPanelTitle}`}
+                >
+                  <TutorialViewComponent
+                    tutorials={this.state.tutorials}
+                    tutorialId={this.state.tutorialId}
+                    stepId={this.state.stepId}
+                    onTutorialContentLoaded={this.handleLoadedTutorial}
+                    onNavigationRequest={this.handleNavigationRequest}
+                    onFixTheCode={this.fixTheCode}
+                  />
+                </ReactGoldenLayoutPanel>
+              </ReactGoldenLayout>
             )}
         </div>
       </div>
