@@ -2,11 +2,9 @@ import * as React from "react";
 import * as cn from "classnames";
 import * as FileSaver from "file-saver";
 import { RouteComponentProps } from "react-router-dom";
-import { Subject } from "rxjs/Subject";
-import { BehaviorSubject } from "rxjs/BehaviorSubject";
+import { BehaviorSubject } from "rxjs";
 
-import { RandomHelper } from "app/utils/random-helper";
-import { callActionSafe, ErrorDef } from "app/utils/error-helpers";
+import { callActionSafe } from "app/utils/error-helpers";
 import { ensure } from "app/utils/syntax-helpers";
 import { resolveInject } from "app/di";
 import { $T } from "app/i18n/strings";
@@ -28,7 +26,7 @@ import { SignInStatusComponent } from "app/ui/sign-in-status.component";
 import { SimpleSelectComponent } from "app/ui/_generic/simple-select.component";
 import { PageHeaderComponent } from "app/ui/_generic/page-header.component";
 import { MainMenuComponent } from "app/ui/main-menu.component";
-import { LogoExecutorComponent } from "app/ui/_shared/logo-executor.component";
+import { LogoExecutorComponent } from "app/ui/_generic/logo-executor/logo-executor.component";
 import { FileSelectorComponent } from "app/ui/_generic/file-selector.component";
 import { LogoCodeSamplesService } from "app/services/program/logo-code-samples.service";
 
@@ -45,7 +43,6 @@ interface IComponentState {
   currentLocale?: ILocaleInfo;
   theme?: Theme;
   userSettings?: IUserSettings;
-  turtleImage?: HTMLImageElement;
 }
 
 interface IComponentProps extends RouteComponentProps<void> {}
@@ -64,8 +61,7 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
   private demoSamplesService = resolveInject(LogoCodeSamplesService);
   private diSetup = resolveInject(DependecyInjectionSetupService);
 
-  private onIsRunningChanged = new Subject<boolean>();
-  private runCode = new BehaviorSubject<string>("");
+  private codeRunnerSubject = new BehaviorSubject<string>("");
   private exportInportService = new PersonalGalleryImportService();
 
   constructor(props: IComponentProps) {
@@ -93,19 +89,22 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
     if (programs && userSettings) {
       const theme = this.themeService.getActiveTheme() || this.themeService.getTheme("Default");
       const locale = this.localizationService.getLocaleById(userSettings.localeId);
-      const turtleImage = this.turtleCustomizationService.getTurtleImage(userSettings.turtleId);
 
-      this.setState({
-        userSettings: userSettings,
-        theme: theme,
-        currentLocale: locale,
-        turtleImage: turtleImage,
-        programCount: programs.length
-      });
-
-      this.runCode.next(this.demoSamplesService.getRandomSample());
+      this.setState(
+        {
+          userSettings: userSettings,
+          theme: theme,
+          currentLocale: locale,
+          programCount: programs.length
+        },
+        this.runRandomProgram
+      );
     }
   }
+
+  private runRandomProgram = () => {
+    this.codeRunnerSubject.next(this.demoSamplesService.getRandomSample());
+  };
 
   private doExport = async () => {
     const programs = await callActionSafe(this.errorService.handleError, async () => this.galleryService.getAll());
@@ -172,7 +171,9 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                                 idAttr="language-selector"
                                 selectionChanged={async selectedLocation => {
                                   if (selectedLocation) {
-                                    await this.userSettingsService.update({ localeId: selectedLocation.id });
+                                    await this.userSettingsService.update({
+                                      localeId: selectedLocation.id
+                                    });
                                     await this.diSetup.reset();
                                   }
                                 }}
@@ -198,9 +199,11 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                                 selectionChanged={async selectedTheme => {
                                   if (selectedTheme) {
                                     this.setState({ theme: selectedTheme });
-                                    await this.userSettingsService.update({ themeName: selectedTheme.name });
+                                    await this.userSettingsService.update({
+                                      themeName: selectedTheme.name
+                                    });
                                     this.themeService.setActiveTheme(selectedTheme.name);
-                                    await this.loadData();
+                                    this.forceUpdate(this.runRandomProgram);
                                   }
                                 }}
                               />
@@ -224,7 +227,7 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                                 selectionChanged={async newTurtle => {
                                   if (newTurtle) {
                                     await this.userSettingsService.update({ turtleId: newTurtle.id });
-                                    await this.loadData();
+                                    this.forceUpdate(this.runRandomProgram);
                                   }
                                 }}
                               />
@@ -248,7 +251,7 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                                 selectionChanged={async newSize => {
                                   if (newSize) {
                                     await this.userSettingsService.update({ turtleSize: newSize.size });
-                                    await this.loadData();
+                                    this.forceUpdate(this.runRandomProgram);
                                   }
                                 }}
                               />
@@ -281,17 +284,12 @@ export class UserProfilePageComponent extends React.Component<IComponentProps, I
                   <div className="column">
                     <div className="card">
                       <div className="card-content">
-                        {[
-                          <LogoExecutorComponent
-                            key={`${JSON.stringify(this.state.userSettings)}`} //this is a hack to force component to be created each render in order to not handle prop change event
-                            onIsRunningChanged={this.onIsRunningChanged}
-                            runCommands={this.runCode}
-                            stopCommands={new Subject<void>()}
-                            customTurtleImage={this.state.turtleImage}
-                            customTurtleSize={this.state.userSettings.turtleSize}
-                            isDarkTheme={this.state.theme.isDark}
-                          />
-                        ]}
+                        <LogoExecutorComponent
+                          runCommands={this.codeRunnerSubject}
+                          isDarkTheme={this.state.theme.isDark}
+                          turtleImage={this.turtleCustomizationService.getTurtleImage(this.state.userSettings.turtleId)}
+                          turtleSize={this.state.userSettings.turtleSize}
+                        />
                       </div>
                     </div>
                   </div>
