@@ -30,14 +30,14 @@ export class Lazy<T> {
 }
 
 export type DictionaryLike<V> = { [name: string]: V };
+export type ValueType = string | number | boolean | Date | undefined;
 
-export function as<T>(object: T): T {
-  return object;
-}
+// tslint:disable-next-line:no-null-keyword
+export const NULL = null;
 
 export interface SortingOptions<T> {
   direction?: "desc" | "asc";
-  sortBy(item: T): string | number | boolean | Date | undefined;
+  sortBy(item: T): ValueType;
 }
 
 /**
@@ -45,6 +45,7 @@ export interface SortingOptions<T> {
  * comparison for strings
  * @param criterions
  */
+// tslint:disable-next-line:cognitive-complexity
 export function createCompareFunction<T>(criterions: SortingOptions<T>[]) {
   return function(itemA: T, itemB: T) {
     for (const options of criterions) {
